+++ log4j1-appenders/src/test/java/com/kdgregory/log4j/aws/TestJsonLayout.java
@@ -37,6 +37,8 @@ import org.apache.log4j.WriterAppender;
 import org.apache.log4j.helpers.LogLog;
 
 import net.sf.kdgcommons.lang.StringUtil;
+import static net.sf.kdgcommons.test.StringAsserts.*;
+
 import net.sf.practicalxml.converter.JsonConverter;
 import net.sf.practicalxml.junit.DomAsserts;
 import net.sf.practicalxml.xpath.XPathWrapper;
@@ -60,7 +62,7 @@ public class TestJsonLayout
     private void initialize(String propsName)
     throws Exception
     {
-        URL config = ClassLoader.getSystemResource(propsName);
+        URL config = ClassLoader.getSystemResource("TestJsonLayout/" + propsName + ".properties");
         PropertyConfigurator.configure(config);
 
         logger = Logger.getLogger(getClass());
@@ -141,7 +143,7 @@ public class TestJsonLayout
     @Test
     public void testSimpleMessage() throws Exception
     {
-        initialize("TestJsonLayout/default.properties");
+        initialize("default");
 
         logger.debug(TEST_MESSAGE);
 
@@ -163,7 +165,7 @@ public class TestJsonLayout
     @Test
     public void testException() throws Exception
     {
-        initialize("TestJsonLayout/default.properties");
+        initialize("default");
 
         String innerMessage = "I'm not worthy";
         String outerMessage = "throw it out";
@@ -200,7 +202,7 @@ public class TestJsonLayout
     @Test
     public void testNDC() throws Exception
     {
-        initialize("TestJsonLayout/default.properties");
+        initialize("default");
 
         NDC.push("frist");  // misspelling intentional
         NDC.push("second");
@@ -218,7 +220,7 @@ public class TestJsonLayout
     @Test
     public void testMDC() throws Exception
     {
-        initialize("TestJsonLayout/default.properties");
+        initialize("default");
 
         MDC.put("foo", "bar");
         MDC.put("argle", "bargle");
@@ -239,7 +241,7 @@ public class TestJsonLayout
     @Test
     public void testLocation() throws Exception
     {
-        initialize("TestJsonLayout/testLocation.properties");
+        initialize("testLocation");
 
         logger.debug(TEST_MESSAGE);
 
@@ -259,7 +261,7 @@ public class TestJsonLayout
     @Test
     public void testDisableHostname() throws Exception
     {
-        initialize("TestJsonLayout/testDisableHostname.properties");
+        initialize("testDisableHostname");
 
         logger.debug(TEST_MESSAGE);
 
@@ -274,7 +276,7 @@ public class TestJsonLayout
     @Ignore("this test should only be run on an EC2 instance")
     public void testInstanceId() throws Exception
     {
-        initialize("TestJsonLayout/testInstanceId.properties");
+        initialize("testInstanceId");
 
         logger.debug(TEST_MESSAGE);
 
@@ -288,9 +290,25 @@ public class TestJsonLayout
 
 
     @Test
+    @Ignore("this test should only be run if you have AWS credentials")
+    public void testAccountId() throws Exception
+    {
+        initialize("testAccountId");
+
+        logger.debug(TEST_MESSAGE);
+
+        captureLoggingOutputAndParse();
+        assertCommonElements(TEST_MESSAGE);
+
+        String accountId = new XPathWrapper("/data/accountId").evaluateAsString(dom);
+        assertRegex("\\d{12}", accountId);
+    }
+
+
+    @Test
     public void testTags() throws Exception
     {
-        initialize("TestJsonLayout/testTags.properties");
+        initialize("testTags");
 
         logger.debug(TEST_MESSAGE);
 
@@ -308,7 +326,7 @@ public class TestJsonLayout
     @Test
     public void testEmptyTags() throws Exception
     {
-        initialize("TestJsonLayout/testEmptyTags.properties");
+        initialize("testEmptyTags");
 
         logger.debug(TEST_MESSAGE);
 
@@ -322,7 +340,7 @@ public class TestJsonLayout
     @Test
     public void testNoAppendNewlines() throws Exception
     {
-        initialize("TestJsonLayout/default.properties");
+        initialize("default");
 
         logger.debug(TEST_MESSAGE);
         logger.debug(TEST_MESSAGE);
@@ -336,7 +354,7 @@ public class TestJsonLayout
     @Test
     public void testAppendNewlines() throws Exception
     {
-        initialize("TestJsonLayout/testAppendNewlines.properties");
+        initialize("testAppendNewlines");
 
         logger.debug(TEST_MESSAGE);
         logger.debug(TEST_MESSAGE);